readme: remove rules and philosophy
Over time I figured out that I could not make a library that could do depend and at the same time don't depend on itself. The problem is that many implementations happen to be usable by other functions and combining multiple of those features together is not way unavoidable. Although, I can't keep with the original rules and philosophy, I'll came up with a better
concept.